[PAN-2422] Add in-memory cache over world state download queue (#1087)
