+++ ethereum/eth/src/test/java/tech/pegasys/pantheon/ethereum/eth/sync/worldstate/WorldDownloadStateTest.java
@@ -28,8 +28,9 @@ import tech.pegasys.pantheon.ethereum.eth.manager.task.EthTask;
 import tech.pegasys.pantheon.ethereum.storage.keyvalue.KeyValueStorageWorldStateStorage;
 import tech.pegasys.pantheon.ethereum.worldstate.WorldStateStorage;
 import tech.pegasys.pantheon.services.kvstore.InMemoryKeyValueStorage;
-import tech.pegasys.pantheon.services.queue.InMemoryTaskQueue;
-import tech.pegasys.pantheon.services.queue.TaskQueue.Task;
+import tech.pegasys.pantheon.services.tasks.CachingTaskCollection;
+import tech.pegasys.pantheon.services.tasks.InMemoryTaskQueue;
+import tech.pegasys.pantheon.services.tasks.Task;
 import tech.pegasys.pantheon.util.bytes.BytesValue;
 
 import java.util.Arrays;
@@ -51,7 +52,8 @@ public class WorldDownloadStateTest {
 
   private final BlockHeader header =
       new BlockHeaderTestFixture().stateRoot(ROOT_NODE_HASH).buildHeader();
-  private final InMemoryTaskQueue<NodeDataRequest> pendingRequests = new InMemoryTaskQueue<>();
+  private final CachingTaskCollection<NodeDataRequest> pendingRequests =
+      new CachingTaskCollection<>(new InMemoryTaskQueue<>());
   private final ArrayBlockingQueue<Task<NodeDataRequest>> requestsToPersist =
       new ArrayBlockingQueue<>(100);
 
@@ -94,7 +96,7 @@ public class WorldDownloadStateTest {
 
   @Test
   public void shouldNotCompleteWhenThereArePendingTasks() {
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
 
     downloadState.checkCompletion(worldStateStorage, header);
 
@@ -115,8 +117,8 @@ public class WorldDownloadStateTest {
     downloadState.addOutstandingTask(outstandingTask1);
     downloadState.addOutstandingTask(outstandingTask2);
 
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY));
     requestsToPersist.add(toPersist1);
     requestsToPersist.add(toPersist2);
 
@@ -133,7 +135,7 @@ public class WorldDownloadStateTest {
 
   @Test
   public void shouldNotSendAdditionalRequestsWhenWaitingForANewPeer() {
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
 
     downloadState.setWaitingForNewPeer(true);
     downloadState.whileAdditionalRequestsCanBeSent(mustNotBeCalled());
@@ -141,7 +143,7 @@ public class WorldDownloadStateTest {
 
   @Test
   public void shouldResumeSendingAdditionalRequestsWhenNoLongerWaitingForPeer() {
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
     final Runnable sendRequest =
         mockWithAction(() -> downloadState.addOutstandingTask(mock(EthTask.class)));
 
@@ -155,10 +157,10 @@ public class WorldDownloadStateTest {
 
   @Test
   public void shouldStopSendingAdditionalRequestsWhenPendingRequestsIsEmpty() {
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
 
-    final Runnable sendRequest = mockWithAction(pendingRequests::dequeue);
+    final Runnable sendRequest = mockWithAction(pendingRequests::remove);
     downloadState.whileAdditionalRequestsCanBeSent(sendRequest);
 
     verify(sendRequest, times(2)).run();
@@ -166,7 +168,7 @@ public class WorldDownloadStateTest {
 
   @Test
   public void shouldStopSendingAdditionalRequestsWhenMaximumOutstandingRequestCountReached() {
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
     final Runnable sendRequest =
         mockWithAction(() -> downloadState.addOutstandingTask(mock(EthTask.class)));
 
@@ -176,7 +178,7 @@ public class WorldDownloadStateTest {
 
   @Test
   public void shouldStopSendingAdditionalRequestsWhenFutureIsCancelled() {
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
     final Runnable sendRequest = mockWithAction(() -> future.cancel(true));
 
     downloadState.whileAdditionalRequestsCanBeSent(sendRequest);
@@ -185,7 +187,7 @@ public class WorldDownloadStateTest {
 
   @Test
   public void shouldStopSendingAdditionalRequestsWhenDownloadIsMarkedAsStalled() {
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
     final Runnable sendRequest = mockWithAction(() -> downloadState.requestComplete(false));
 
     downloadState.whileAdditionalRequestsCanBeSent(sendRequest);
@@ -210,7 +212,7 @@ public class WorldDownloadStateTest {
 
   @Test
   public void shouldNotAllowMultipleCallsToSendAdditionalRequestsAtOnce() {
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
     final Runnable sendRequest =
         mockWithAction(
             () -> {
@@ -223,7 +225,7 @@ public class WorldDownloadStateTest {
   }
 
   @Test
-  public void shouldNotEnqueueRequestsAfterDownloadIsStalled() {
+  public void shouldNotAddRequestsAfterDownloadIsStalled() {
     downloadState.checkCompletion(worldStateStorage, header);
 
     downloadState.enqueueRequests(Arrays.asList(createAccountDataRequest(Hash.EMPTY_TRIE_HASH)));
@@ -235,7 +237,7 @@ public class WorldDownloadStateTest {
   @Test // Sanity check for the test structure
   public void shouldFailWhenMustNotBeCalledIsCalled() {
 
-    pendingRequests.enqueue(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
+    pendingRequests.add(createAccountDataRequest(Hash.EMPTY_TRIE_HASH));
     assertThatThrownBy(() -> downloadState.whileAdditionalRequestsCanBeSent(mustNotBeCalled()))
         .hasMessage("Unexpected invocation");
   }
