+++ services/tasks/src/main/java/tech/pegasys/pantheon/services/tasks/TaskCollection.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2019 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package tech.pegasys.pantheon.services.tasks;
+
+import java.io.Closeable;
+
+public interface TaskCollection<T> extends Closeable {
+  /**
+   * Add some data that needs to be processed.
+   *
+   * @param taskData The data to be processed.
+   */
+  void add(T taskData);
+
+  /**
+   * Get a task for processing. This task will be tracked as a pending task until either {@code
+   * Task.markCompleted} or {@code Task.requeue} is called.
+   *
+   * @return The task to be processed.
+   */
+  Task<T> remove();
+
+  /** @return The number of tasks in the queue. */
+  long size();
+
+  /** @return True if all tasks have been dequeued. */
+  boolean isEmpty();
+
+  /** Clear all data from the queue. */
+  void clear();
+
+  /** @return True if all tasks have been dequeued and processed. */
+  boolean allTasksCompleted();
+}
