+++ settings.gradle
@@ -36,6 +36,7 @@ include 'metrics'
 include 'pantheon'
 include 'services:kvstore'
 include 'services:pipeline'
+include 'services:tasks'
 include 'services:queue'
 include 'services:util'
 include 'testutil'
